Introduce new single-page design and teal theme
First and foremost, this includes a new single-page design. All of the
content is now on index.html. This should make navigating easier.

Instead of the white colors, there is now a teal theme.

There is a custom nav-bar with some JS to expand and close the menu
on mobile.